enble float64 and better tc for pt2

+++ ad_afqmc/prop_unrestricted/run_afqmc_ccsd_pt.py
@@ -77,19 +77,19 @@ for n in range(1,options["n_eql"]+1):
         sample_ccsd_pt.propagate_phaseless(
             prop_data, ham_data, prop, trial, wave_data, sampler)
 
-    blk_wt = np.array([blk_wt], dtype="float32") 
-    blk_t = np.array([blk_t], dtype="float32")
-    blk_e0 = np.array([blk_e0], dtype="float32")
-    blk_e1 = np.array([blk_e1], dtype="float32")
+    blk_wt = np.array([blk_wt], dtype="float64") 
+    blk_t = np.array([blk_t], dtype="float64")
+    blk_e0 = np.array([blk_e0], dtype="float64")
+    blk_e1 = np.array([blk_e1], dtype="float64")
 
-    blk_wt_t = np.array([blk_t * blk_wt], dtype="float32")
-    blk_wt_e0 = np.array([blk_e0 * blk_wt], dtype="float32")
-    blk_wt_e1 = np.array([blk_e1 * blk_wt], dtype="float32")
+    blk_wt_t = np.array([blk_t * blk_wt], dtype="float64")
+    blk_wt_e0 = np.array([blk_e0 * blk_wt], dtype="float64")
+    blk_wt_e1 = np.array([blk_e1 * blk_wt], dtype="float64")
 
-    tot_blk_wt = np.zeros(1, dtype="float32")
-    tot_blk_t = np.zeros(1, dtype="float32")
-    tot_blk_e0 = np.zeros(1, dtype="float32")
-    tot_blk_e1 = np.zeros(1, dtype="float32")
+    tot_blk_wt = np.zeros(1, dtype="float64")
+    tot_blk_t = np.zeros(1, dtype="float64")
+    tot_blk_e0 = np.zeros(1, dtype="float64")
+    tot_blk_e1 = np.zeros(1, dtype="float64")
 
     comm.Reduce(
         [blk_wt, MPI.FLOAT],
@@ -151,9 +151,6 @@ comm.Barrier()
 if rank == 0:
     print("#\n# Sampling sweeps:")
     print("#  Iter \t "
-        #   "   <t> \t error \t \t"
-        #   "   <e0> \t error \t \t "
-        #   "   <e1> \t error \t \t"
           "   energy \t error \t \t Walltime")
 comm.Barrier()
 
@@ -164,11 +161,11 @@ glb_blk_e1 = None
 
 comm.Barrier()
 if rank == 0:
-    glb_blk_wt = np.zeros(size * sampler.n_blocks,dtype="float32")
-    glb_blk_t = np.zeros(size * sampler.n_blocks,dtype="float32")
-    glb_blk_e0 = np.zeros(size * sampler.n_blocks,dtype="float32")
-    glb_blk_e1 = np.zeros(size * sampler.n_blocks,dtype="float32")
-    ept_samples = np.zeros(sampler.n_blocks,dtype="float32")
+    glb_blk_wt = np.zeros(size * sampler.n_blocks,dtype="float64")
+    glb_blk_t = np.zeros(size * sampler.n_blocks,dtype="float64")
+    glb_blk_e0 = np.zeros(size * sampler.n_blocks,dtype="float64")
+    glb_blk_e1 = np.zeros(size * sampler.n_blocks,dtype="float64")
+    ept_samples = np.zeros(sampler.n_blocks,dtype="float64")
 comm.Barrier()
     
 for n in range(sampler.n_blocks):
@@ -176,10 +173,10 @@ for n in range(sampler.n_blocks):
         sample_ccsd_pt.propagate_phaseless(
             prop_data, ham_data, prop, trial, wave_data, sampler)
     
-    blk_wt = np.array([blk_wt], dtype="float32")
-    blk_t = np.array([blk_t], dtype="float32")
-    blk_e0 = np.array([blk_e0], dtype="float32")
-    blk_e1 = np.array([blk_e1], dtype="float32")
+    blk_wt = np.array([blk_wt], dtype="float64")
+    blk_t = np.array([blk_t], dtype="float64")
+    blk_e0 = np.array([blk_e0], dtype="float64")
+    blk_e1 = np.array([blk_e1], dtype="float64")
 
     gather_wt = None
     gather_t = None
@@ -188,10 +185,10 @@ for n in range(sampler.n_blocks):
 
     comm.Barrier()
     if rank == 0:
-        gather_wt = np.zeros(size, dtype="float32")
-        gather_t = np.zeros(size, dtype="float32")
-        gather_e0 = np.zeros(size, dtype="float32")
-        gather_e1 = np.zeros(size, dtype="float32")
+        gather_wt = np.zeros(size, dtype="float64")
+        gather_t = np.zeros(size, dtype="float64")
+        gather_e0 = np.zeros(size, dtype="float64")
+        gather_e1 = np.zeros(size, dtype="float64")
     comm.Barrier()
 
     comm.Gather(blk_wt, gather_wt, root=0)
